+++ Code/Engine/Renderer/D3D11/Resources/D3D11ConstantBuffer.cpp
@@ -2,7 +2,7 @@
 
 
 //---------------------------------------------------------------------------------------------------------------------------
-D3D11ConstantBuffer::D3D11ConstantBuffer(size_t byteSizeOfBuffer) 
+D3D11ConstantBuffer::D3D11ConstantBuffer(uint byteSizeOfBuffer)
 	: m_bufferSize(byteSizeOfBuffer)
 {
 	m_pByteBuffer = new byte[byteSizeOfBuffer];
